requirements.txt
attrs==22.1.0
# Editable Git install with no remote (bookmarks==1.0.0)
-e /home/dreamer20/dev/bookmarks
click==8.1.3
coverage==6.5.0
exceptiongroup==1.0.2
Flask==2.2.2
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytest==7.2.0
tomli==2.0.1
Werkzeug==2.2.2
